Reorganize classes and files in different packages
