+++ app/src/main/java/com/androidestudos/fiapchallange/network/APIServerDataSource.kt
@@ -1,9 +1,8 @@
-package com.androidestudos.fiapchallange
+package com.androidestudos.fiapchallange.network
 
 import android.util.Log
+import com.androidestudos.fiapchallange.data.CreateTarefaResult
 import okhttp3.OkHttpClient
-import retrofit2.HttpException
-import retrofit2.Response
 import retrofit2.Retrofit
 import retrofit2.converter.gson.GsonConverterFactory
